# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger",
        "description": "API CRUD Tournament",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://tournament.test"
        }
    ],
    "paths": {
        "/api/player": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get list of players",
                "description": "Returns a list of players",
                "operationId": "86f97704ab21be5419b8fe200540eeff",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PlayerResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Create a new player",
                "description": "Creates a new player and returns the created player data",
                "operationId": "9265cefd0e53f2eb5fce951bf96985d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlayerResource"
                            },
                            "example": {
                                "name": "Jennifer",
                                "gender": "woman",
                                "skill_level": 63,
                                "strength": null,
                                "speed": null,
                                "reaction_time": 74,
                                "tournament_id": 2
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Player created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlayerResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Invalid input"
                            }
                        }
                    }
                }
            }
        },
        "/api/players/{player}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get player by Id",
                "description": "Returns a player by Id",
                "operationId": "1f0417c2ce63e08c2bc18eb1dec0070a",
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlayerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Player not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "Delete player by Id",
                "description": "Returns boolean: true or false",
                "operationId": "65491a560bcdda23bfd292fa1f869317",
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Player not found"
                    }
                }
            }
        },
        "/api/player/{player}": {
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "Player update",
                "operationId": "345e9112a93362524efcc8545573d41a",
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlayerResource"
                            },
                            "example": {
                                "name": "Jennifer",
                                "gender": "woman",
                                "skill_level": 78,
                                "strength": null,
                                "speed": null,
                                "reaction_time": 66,
                                "tournament_id": 2
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Player updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlayerResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Invalid input"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PlayerResource": {
                "required": [
                    "name",
                    "gender",
                    "skill_level",
                    "tournament_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the Player",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the Player",
                        "type": "string",
                        "example": "Matt"
                    },
                    "gender": {
                        "description": "Gender of the Player",
                        "type": "string",
                        "enum": [
                            "man",
                            "woman"
                        ],
                        "example": "man"
                    },
                    "skill_level": {
                        "description": "Player's skill level",
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 60
                    },
                    "strength": {
                        "description": "Player's strength",
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 35
                    },
                    "speed": {
                        "description": "Player's speed",
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 45
                    },
                    "reaction_time": {
                        "description": "Player's reaction time",
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0,
                        "example": null
                    },
                    "tournament_id": {
                        "description": "Tournament's ID",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}